refactor: create base batch class
